# backend/app/models/project.py
from sqlalchemy import Column, String, DateTime, JSON, Enum as SQLEnum
from sqlalchemy.dialects.postgresql import UUID, ARRAY
from sqlalchemy.orm import relationship
from sqlalchemy.sql import func
import uuid
import enum
from app.core.database import Base


class ProjectStatus(str, enum.Enum):
    """Project status enumeration."""
    ACTIVE = "active"
    ARCHIVED = "archived"


class Project(Base):
    """Project model."""

    __tablename__ = "projects"

    id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
    name = Column(String, nullable=False)
    description = Column(String)
    repository_url = Column(String, nullable=False)
    tech_stack = Column(ARRAY(String))
    status = Column(SQLEnum(ProjectStatus), default=ProjectStatus.ACTIVE)
    created_by = Column(UUID(as_uuid=True), nullable=False)
    created_at = Column(DateTime(timezone=True), server_default=func.now())
    updated_at = Column(DateTime(timezone=True), onupdate=func.now())

    # Relationships
    created_by_user = relationship("User", back_populates="projects")
    ai_activities = relationship("AIActivity", back_populates="project")
    agent_executions = relationship("AgentExecution", back_populates="project")
    pipeline_executions = relationship("PipelineExecution", back_populates="project")