# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI-Assisted Developer Productivity Platform API
  description: |
    API for managing AI-assisted development workflows, tracking AI tool usage,
    orchestrating code agents, and integrating with CI/CD pipelines.
  version: 1.0.0
  contact:
    name: API Support
    email: support@ai-dev-platform.example
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://ai-dev-platform-staging.example.com/api/v1
    description: Staging server
  - url: https://ai-dev-platform.example.com/api/v1
    description: Production server
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List all projects
      description: Retrieve a paginated list of projects
      operationId: listProjects
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SearchParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [active, archived]
          description: Filter by status
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Projects]
      summary: Create project
      description: Create a new project
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project
      description: Retrieve a specific project by ID
      operationId: getProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Projects]
      summary: Update project
      description: Update a specific project
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Projects]
      summary: Delete project
      description: Delete a specific project
      operationId: deleteProject
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '204':
          description: Project deleted successfully
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    PerPageParam:
      in: query
      name: per_page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of items per page
    SearchParam:
      in: query
      name: search
      schema:
        type: string
      description: Search term for filtering results
    ProjectIdParam:
      in: path
      name: project_id
      required: true
      schema:
        type: string
        format: uuid
      description: Project UUID
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: object
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        repository_url:
          type: string
          format: uri
        tech_stack:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, archived]
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectCreate:
      type: object
      required: [name, repository_url]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 500
        repository_url:
          type: string
          format: uri
        tech_stack:
          type: array
          items:
            type: string
          maxItems: 20
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 500
        repository_url:
          type: string
          format: uri
        tech_stack:
          type: array
          items:
            type: string
          maxItems: 20
        status:
          type: string
          enum: [active, archived]
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer